+++ previews/elements/buttons_component_preview.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+class Elements::ButtonsComponentPreview < ViewComponent::Preview
+  def default
+  end
+
+  def attached
+  end
+
+  def alignment
+  end
+end