+++ previews/elements/box_component_preview.rb
@@ -2,8 +2,5 @@
 
 class Elements::BoxComponentPreview < ViewComponent::Preview
   def with_text_content
-    render(Bulma::BoxComponent.new) do
-      "I'm in a box."
-    end
   end
 end