+++ previews/elements/notification_component_preview.rb
@@ -3,4 +3,7 @@
 class Elements::NotificationComponentPreview < ViewComponent::Preview
   def default
   end
+
+  def colors
+  end
 end