+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Converted preview for `Bulma::BoxComponent` to use view file
 - Converted default preview for `Bulma::ProgressBarComponent` to use view file
 - Added support to specify content as first arggument to `Bulma::TitleComponent`
 - Drop supoort for Ruby 2,7